refactor(town): extract Wild Seppo logic into seppo_runtime.js and call tickSeppo(ctx) from TownRuntime.tick
